Move example service to new directory structure
