+++ src/assets/thirdPartyServices/snet/example_service/example_service_pb.js
